+++ src/SixLabors.Fonts/Tables/AdvancedTypographic/GSub/LookupType1SubTable.cs
@@ -102,7 +102,7 @@ internal sealed class LookupType1Format1SubTable : LookupSubTable
         GSubTable table,
         ShapingBuffer buffer,
         Tag feature,
-        ulong lookupMask,
+        uint lookupMask,
         int index,
         int count)
     {
@@ -197,7 +197,7 @@ internal sealed class LookupType1Format2SubTable : LookupSubTable
         GSubTable table,
         ShapingBuffer buffer,
         Tag feature,
-        ulong lookupMask,
+        uint lookupMask,
         int index,
         int count)
     {
