+++ src/SixLabors.Fonts/Tables/AdvancedTypographic/GSub/LookupType2SubTable.cs
@@ -124,7 +124,7 @@ internal sealed class LookupType2Format1SubTable : LookupSubTable
         GSubTable table,
         ShapingBuffer buffer,
         Tag feature,
-        ulong lookupMask,
+        uint lookupMask,
         int index,
         int count)
     {
