+++ src/SixLabors.Fonts/Tables/AdvancedTypographic/GSub/LookupType3SubTable.cs
@@ -123,7 +123,7 @@ internal sealed class LookupType3Format1SubTable : LookupSubTable
         GSubTable table,
         ShapingBuffer buffer,
         Tag feature,
-        ulong lookupMask,
+        uint lookupMask,
         int index,
         int count)
     {
